move yxi up a directory
